requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
google-genai>=1.0.0
faiss-cpu>=1.9.0
numpy>=1.26.0
python-dotenv>=1.0.0
httpx>=0.27.0
